+++ packages/svelte/src/lib/components/CallOnMount/index.ts
@@ -1,0 +1,1 @@
+export { default as CallOnMount } from './CallOnMount.svelte';